+++ parallel-first-implementation/e_step.h
@@ -2,9 +2,9 @@
 #define EM_PROJECT_E_STEP_H
 
 // run expectation step
-void e_step(float *X, float *mean, float *cov, float *weights, float *p_val, int K, int N, int D);
+void e_step(double *X, double *mean, double *cov, double *weights, double *p_val, int K, int N, int D);
 
 // computes pdf
-float gaussian(float *x, float *mean, float *cov, int D);
+double gaussian(double *x, double *mean, double *cov, int D);
 
 #endif //EM_PROJECT_E_STEP_H
