+++ parallel-first-implementation/reader.h
@@ -2,7 +2,7 @@
 #define EM_PROJECT_READER_H
 
 // populate the matrix of samples
-void fill_matrix(float* mat, int N, int D, char *FILE_PATH);
+void fill_matrix(double* mat, int N, int D, char *FILE_PATH);
 
 
 #endif //EM_PROJECT_READER_H
