+++ parallel-first-implementation/utils.h
@@ -2,13 +2,13 @@
 #define EM_PROJECT_UTILS_H
 
 //copy values of mean and covariance
-void get_cluster_mean_cov(float *mean, float *cov, float *m_res, float *cov_res, int k, int D);
+void get_cluster_mean_cov(double *mean, double *cov, double *m_res, double *cov_res, int k, int D);
 
 // divide the matrix for each process
 void divide_rows(int* data_count, int* data_displ, int* p_count, int* p_displ,
                  int N, int D, int K, int comm_sz);
 
-void free_em_data(float *X, float *mean, float *cov, float *weights, float *p_val);
+void free_em_data(double *X, double *mean, double *cov, double *weights, double *p_val);
 
 void free_rows_data(int* data_count, int* data_displ, int* p_count, int* p_displ);
